requirements.txt
colorama
matplotlib
numpy
opencv_python
Pillow
PyYAML
tqdm
ultralytics
streamlit
wandb
codecarbon
plotly